Scrapers/youm7/requirements.txt
boto3
botocore
python-dotenv
beautifulsoup4
lxml
selenium
webdriver-manager
urllib3
requests